+++ packages/api/src/page/page.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common'
 import { TypeOrmModule } from '@nestjs/typeorm'
-import { Page, PageRepository, PageService } from '@webok/core/lib/page'
+import { Page } from '@webok/core/lib/page'
+import { PageRepository, PageService } from '@webok/services/lib/page'
 import { PageController } from './page.controller'
 
 @Module({
   imports: [TypeOrmModule.forFeature([Page, PageRepository])],
-  providers: [PageService],
+  providers: [{ provide: 'IPageService', useClass: PageService }],
   controllers: [PageController],
 })
 export class PageModule {}
