+++ packages/api/src/typeorm/index.ts
@@ -1,0 +1,1 @@
+export * from './typeorm.module'
