+++ packages/api/src/typeorm/typeorm.module.ts
@@ -1,4 +1,4 @@
 import { TypeOrmModule } from '@nestjs/typeorm'
-import TypeOrmConfig from '@webok/core/lib/typeorm.config'
+import TypeOrmConfig from '@webok/core/lib/typeorm/typeorm.config'
 
 export const TypeOrmRootModule = TypeOrmModule.forRoot(TypeOrmConfig)
