+++ packages/api/src/user/user.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common'
 import { TypeOrmModule } from '@nestjs/typeorm'
-import { User, UserRepository, UserService } from '@webok/core/lib/user'
+import { User } from '@webok/core/lib/user'
+import { UserRepository, UserService } from '@webok/services/lib/user'
 import { UserController } from './user.controller'
 
 @Module({
   imports: [TypeOrmModule.forFeature([User, UserRepository])],
-  providers: [UserService],
+  providers: [{ provide: 'IUserService', useClass: UserService }],
   controllers: [UserController],
 })
 export class UserModule {}
