+++ packages/core/src/page/index.ts
@@ -1,4 +1,3 @@
 export * from './page.entity'
-export * from './page.repository'
 export * from './page.data'
-export * from './page.service'
+export * from './page-service.interface'
