+++ packages/core/src/user/index.ts
@@ -1,4 +1,3 @@
 export * from './user.entity'
-export * from './user.repository'
 export * from './user.data'
-export * from './user.service'
+export * from './user-service.interface'
