+++ packages/core/src/user/user-service.interface.ts
@@ -1,0 +1,6 @@
+import { User } from './user.entity'
+import { CreateUserData } from './user.data'
+
+export interface IUserService {
+  create (data: CreateUserData): Promise<User>
+}
