+++ packages/services/src/page/index.ts
@@ -1,0 +1,2 @@
+export * from './page.repository'
+export * from './page.service'
