+++ packages/services/src/page/page.repository.ts
@@ -1,5 +1,5 @@
 import { Repository, EntityRepository } from 'typeorm'
-import { Page } from './page.entity'
+import { Page } from '@webok/core/lib/page'
 
 @EntityRepository(Page)
 export class PageRepository extends Repository<Page> {}
