+++ packages/services/src/user/user.repository.ts
@@ -1,5 +1,5 @@
 import { Repository, EntityRepository } from 'typeorm'
-import { User } from './user.entity'
+import { User } from '@webok/core/lib/user'
 
 @EntityRepository(User)
 export class UserRepository extends Repository<User> {}
