+++ packages/services/tslint.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./node_modules/@webok/dev/tslint.json"
+}
